GEOS-5 GESO-17 GEOS-18 -Fixed bug that gave -inf values in seismic time. -Included some lines in segy TextualHeader -Changed from double to float in old UTM coordinates for segy files. -Removed bug that did not stack seismic for storm output when no specific stack is asked for.

+++ geo2seis/utils/segy_writer.cpp
@@ -128,8 +128,8 @@ void SEGY::writeSegy(NRLib::StormContGrid      &data,
     for (size_t i = 0; i < nx; i++) {
       xt = double((i + 0.5) * dx);
       yt = double((j + 0.5) * dy);
-      x = double(geometry->GetX0() + xt * geometry->GetCosRot() - yt * geometry->GetSinRot());
-      y = double(geometry->GetY0() + yt * geometry->GetCosRot() + xt * geometry->GetSinRot());
+      x = static_cast<float>(geometry->GetX0() + xt * geometry->GetCosRot() - yt * geometry->GetSinRot());
+      y = static_cast<float>(geometry->GetY0() + yt * geometry->GetCosRot() + xt * geometry->GetSinRot());
 
       double zbot = data.GetBotSurface().GetZ(x, y);
       double ztop = data.GetTopSurface().GetZ(x, y);
